refactor: final tagging cleanup
In this PR I move tagging related functionality around in order for it to be in a resonable place. Not done before to have the  previous PRs be reviewable.

The sender sync algo was implemented first and for this reason it was in `tagging/sync` dir. In this PR I moved it to `tagging/sender_sync` and move a bunch of other stuff around.
