+++ yarn-project/pxe/src/logs/log_service.test.ts
@@ -12,8 +12,8 @@ import { LogRetrievalRequest } from '../contract_function_simulator/noir-structs
 import { AddressDataProvider } from '../storage/address_data_provider/address_data_provider.js';
 import { AnchorBlockDataProvider } from '../storage/anchor_block_data_provider/anchor_block_data_provider.js';
 import { CapsuleDataProvider } from '../storage/capsule_data_provider/capsule_data_provider.js';
+import { RecipientTaggingDataProvider } from '../storage/tagging_data_provider/recipient_tagging_data_provider.js';
 import { SenderAddressBook } from '../storage/tagging_data_provider/sender_address_book.js';
-import { RecipientTaggingDataProvider } from '../tagging/recipient_sync/recipient_tagging_data_provider.js';
 import { LogService } from './log_service.js';
 
 describe('LogService', () => {
