+++ yarn-project/pxe/src/storage/tagging_data_provider/index.ts
@@ -1,3 +1,3 @@
 export { SenderTaggingDataProvider } from './sender_tagging_data_provider.js';
 export { SenderAddressBook } from './sender_address_book.js';
-export { RecipientTaggingDataProvider } from '../../tagging/recipient_sync/recipient_tagging_data_provider.js';
+export { RecipientTaggingDataProvider } from './recipient_tagging_data_provider.js';
