+++ yarn-project/pxe/src/tagging/sender_sync/utils/load_and_store_new_tagging_indexes.ts
